Rename Guid to guid
